assessment_Project/requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2
django-cors-headers>=4.0
psycopg2-binary>=2.9  # optional if you use Postgres
python-dotenv>=1.0.0
